xdpilone

Struct DeviceQueue

pub struct DeviceQueue { /* private fields */ }

One device queue associated with an XDP socket.

A socket is more specifically a set of receive and transmit queues for packets (mapping to some underlying hardware mapping those bytes with a network). The fill and completion queue can, in theory, be shared with other sockets of the same Umem.

Implementations

impl DeviceQueue

pub fn statistics(&self) -> Result<XdpStatistics, Errno>

👎Deprecated: Consider using statistics_v2 for additional statistics exposed on >= Linux 5.9

Get the statistics of this XDP socket.

pub fn statistics_v2(&self) -> Result<XdpStatisticsV2, Errno>

Get the statistics of this XDP socket.

pub fn bind(&self, interface: &User) -> Result<(), Errno>

Bind the socket to a device queue, activate rx/tx queues.

impl DeviceQueue

pub fn fill(&mut self, max: u32) -> WriteFill<'_>

Prepare some buffers for the fill ring.

The argument is an upper bound of buffers. Use the resulting object to pass specific buffers to the fill queue and commit the write.

pub fn complete(&mut self, n: u32) -> ReadComplete<'_>

Reap some buffers from the completion ring.

Return an iterator over completed buffers.

The argument is an upper bound of buffers. Use the resulting object to dequeue specific buffers from the completion queue and commit the read.

pub fn available(&self) -> u32

Return the difference between our the kernel’s producer state and our consumer head.

pub fn pending(&self) -> u32

Return the difference between our committed consumer state and the kernel’s producer state.

pub fn as_raw_fd(&self) -> c_int

Get the raw file descriptor of this ring.

Safety

Use the file descriptor to attach the ring to an XSK map, for instance, but do not close it and avoid modifying it (unless you know what you’re doing). It should be treated as a BorrowedFd<'_>. That said, it’s not instant UB but probably delayed UB when the DeviceQueue modifies a reused file descriptor that it assumes to own.

pub fn needs_wakeup(&self) -> bool

Query if the fill queue needs to be woken to proceed receiving.

This is only accurate if Umem::XDP_BIND_NEED_WAKEUP was set.

pub fn wake(&mut self)

Poll the fill queue descriptor, to wake it up.