xdpilone

Struct RingTx

pub struct RingTx { /* private fields */ }

A transmitter queue.

This also maintains the mmap of the associated queue.

Implementations

impl RingTx

pub fn transmit(&mut self, n: u32) -> WriteTx<'_>

Transmit some buffers.

Returns a proxy that can be fed descriptors.

pub fn pending(&self) -> u32

Return the difference between our committed producer state and the kernel’s consumer head.

pub fn needs_wakeup(&self) -> bool

Query if the transmit queue needs to be woken to proceed receiving.

This is only accurate if Umem::XDP_BIND_NEED_WAKEUP was set.

pub fn wake(&self)

Send a message (with MSG_DONTWAIT) to wake up the transmit queue.

pub fn as_raw_fd(&self) -> c_int

Get the raw file descriptor of this TX ring.

Safety

Use the file descriptor to attach the ring to an XSK map, for instance, but do not close it and avoid modifying it (unless you know what you’re doing). It should be treated as a BorrowedFd<'_>. That said, it’s not instant UB but probably delayed UB when the RingTx modifies a reused file descriptor that it assumes to own (for instance, wake sends a message to it).