xdpilone

Struct Umem

pub struct Umem { /* private fields */ }

The basic Umem descriptor.

This struct manages the buffers themselves, in a high-level sense, not any of the communication or queues.

Compared to libxdp there no link to the queues is stored. Such a struct would necessitate thread-safe access to the ring’s producer and consumer queues. Instead, a DeviceQueue is the owner of a device queue’s fill/completion ring, but not receive and transmission rings. All other sockets with the same interface/queue depend on it but have their own packet rings.

You’ll note that the fill ring and completion are a shared liveness requirement but under unique control. Exactly one process has to responsibility of maintaining them and ensuring the rings progress. Failing to do so impacts all sockets sharing this Umem. The converse is not true. A single socket can starve its transmission buffer or refuse accepting received packets but the worst is packet loss in this queue.

The controller of the fill/completion pair also controls the associated bpf program which maps packets onto the set of sockets (aka. ‘XSKMAP’).

Implementations

impl Umem

pub unsafe fn new( config: UmemConfig, area: NonNull<[u8]>, ) -> Result<Umem, Errno>

Create a new Umem ring.

Safety

The caller passes an area denoting the memory of the ring. It must be valid for the indicated buffer size and count. The caller is also responsible for keeping the mapping alive.

The area must be page aligned and not exceed i64::MAX in length (on future systems where you could).

pub fn frame(&self, idx: BufIdx) -> Option<UmemChunk>

Get the address associated with a buffer, if it is in-bounds.

Safety

No requirements. However, please ensure that use of the pointer is done properly. The pointer is guaranteed to be derived from the area passed in the constructor. The method guarantees that it does not access any of the pointers in this process.

pub fn len_frames(&self) -> u32

Count the number of available data frames.

pub fn fq_cq(&self, interface: &Socket) -> Result<DeviceQueue, Errno>

Configure the fill and completion queue for a interface queue.

The caller should only call this once for each interface info. However, it’s not entirely incorrect to do it multiple times. Just, be careful that the administration becomes extra messy. All code is written under the assumption that only one controller/writer for the user-space portions of each queue is active at a time. The kernel won’t care about your broken code and race conditions writing to the same queue concurrently. It’s an SPSC. Probably only the first call for each interface succeeds.

pub fn rx_tx( &self, interface: &Socket, config: &SocketConfig, ) -> Result<User, Errno>

Configure the device address for a socket.

Either rx_size or tx_size must be non-zero, i.e. the call to bind will fail if none of the rings is actually configured.

Note: if the underlying socket is shared then this will also bind other objects that share the underlying socket file descriptor, this is intended.

pub fn bind(&self, interface: &User) -> Result<(), Errno>

Activate a socket with by binding it to a device.

This associates the umem region to these queues. This is intended for:

sockets that maintain the fill and completion ring for a device queue, i.e. a fc_cq was called with the socket and that network interface queue is currently being bound.

queues that the umem socket file descriptor is maintaining as a device queue, i.e. the call to fc_cq used a socket created with Socket::with_shared that utilized the Umem instance.

Otherwise, when a pure rx/tx socket should be setup use DeviceQueue::bind with the previously bound socket providing its fill/completion queues.

The tree of parents should look as follows:

fd0: umem [+fq/cq for ifq0] [+rx/+tx]
|- [fd1: socket +rx/tx on ifq0 if fd0 has fq/cq] Umem::bind(fd0, fd1)
|- [fd2: socket +rx/tx on ifq0 if fd0 has fq/cq …] Umem::bind(fd0, fd2)
|
|- fd3: socket +fq/cq for ifq1 [+rx/tx] Umem::bind(fd0, fd3)
| |- fd4: socket +rx/tx on ifq1 DeviceQueue::bind(fd3, fd4)
| |- fd5: socket +rx/tx on ifq1 … DeviceQueue::bind(fd3, fd5)
|
|-fd6:  socket +fq/cq for ifq2 [+rx/tx] Umem::bind(fd0, fd6)
| |- fd7: socket +rx/tx on ifq1 DeviceQueue::bind(fd6, fd7)
| |- …