xdpilone

Struct User

pub struct User { /* private fields */ }

An owner of receive/transmit queues.

This represents a configured version of the raw Socket. It allows you to map the required rings and then Umem::bind the socket, enabling the operations of the queues with the interface.

Implementations

impl User

pub fn statistics(&self) -> Result<XdpStatistics, Errno>

👎Deprecated: Consider using statistics_v2 for additional statistics exposed on >= Linux 5.9

Get the statistics of this XDP socket.

pub fn statistics_v2(&self) -> Result<XdpStatisticsV2, Errno>

Get the statistics of this XDP socket.

pub fn map_rx(&self) -> Result<RingRx, Errno>

Map the RX ring into memory, returning a handle.

Fails if you did not pass any size for rx_size in the configuration, which should be somewhat obvious.

FIXME: we allow mapping the ring more than once. Not a memory safety problem afaik, but a correctness problem.

pub fn map_tx(&self) -> Result<RingTx, Errno>

Map the TX ring into memory, returning a handle.

Fails if you did not pass any size for tx_size in the configuration, which should be somewhat obvious.

FIXME: we allow mapping the ring more than once. Not a memory safety problem afaik, but a correctness problem.

impl User

pub fn as_raw_fd(&self) -> i32

Get the raw file descriptor number underlying this socket.