xdpilone

Struct WriteFill

pub struct WriteFill<'queue> { /* private fields */ }

A writer to a fill queue.

Created with DeviceQueue::fill.

The owner of this value should call some of the insertion methods in any order, then release the writes by WriteFill::commit which performs an atomic release in the Umem queue.

Implementations

impl WriteFill<'_>

pub fn capacity(&self) -> u32

The total number of available slots.

pub fn insert_once(&mut self, nr: u64) -> bool

Fill one device descriptor to be filled.

A descriptor is an offset in the respective Umem’s memory. Any offset within a chunk can be used to mark the chunk as available for fill. The kernel will overwrite the contents arbitrarily until the chunk is returned via the RX queue.

Returns if the insert was successful, that is false if the ring is full. It’s guaranteed that the first WriteFill::capacity inserts with this function succeed.

pub fn insert(&mut self, it: impl Iterator<Item = u64>) -> u32

Fill additional slots that were reserved.

The iterator is polled only for each available slot until either is empty. Returns the total number of slots filled.

pub fn commit(&mut self)

Commit the previously written buffers to the kernel.